Earning with Bigdecimal

+++ demo/src/main/java/com/example/demo/dto/StatDTO.java
@@ -1,22 +1,23 @@
 package com.example.demo.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class StatDTO {
     private String from;
     private String to;
-    private double earning;
+    private BigDecimal earning;
 
     public String getFrom() { return from;}
     public String getTo() { return to;}
-    public double getEarning() { return earning;}
+    public BigDecimal getEarning() { return earning;}
 
     public void setFrom(String from) { this.from = from;}
     public void setTo(String to) { this.to = to;}
-    public void setEarning(double earning) { this.earning = earning;}
+    public void setEarning(BigDecimal earning) { this.earning = earning;}
 
     public StatDTO(){}
-    public StatDTO(String from, String to, double earning) {
+    public StatDTO(String from, String to, BigDecimal earning) {
         this.from = from;
         this.to = to;
         this.earning = earning;
