+++ demo/src/main/java/com/example/demo/entity/Stat.java
@@ -1,19 +1,20 @@
 package com.example.demo.entity;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Stat {
     private Date date;
-    private double earning;
+    private BigDecimal earning;
 
     public Date getDate() {return date;}
-    public double getEarning() {return earning;}
+    public BigDecimal getEarning() {return earning;}
 
     public void setDate(Date date) {this.date = date;}
-    public void setEarning(double earning) {this.earning = earning;}
+    public void setEarning(BigDecimal earning) {this.earning = earning;}
 
     public Stat(){}
-    public Stat(Date date, double earning) {
+    public Stat(Date date, BigDecimal earning) {
         this.date = date;
         this.earning = earning;
     }
